Add macros for registering arbitrary boxed types in user code.
This is helpful if you want to mix the bindings with some native
C code and you need to use a boxed type with either a property
or a signal. Every other type can be mapped to something else
and used through the existing bindings, but there is nothing for
boxed types.