Removed access to deprecated methods
The deprecated method gst_x_overlay_set_xwindow_id has been removed from
GStreamer 0.10.31. This patch aligns the Gst::XOverlay interface with the
gstreamer counterpart and changes the names of the methods to avoid
ambiguity. For better conformance to Qt API, the WId type is also used
for parameters to identify the window handle.