Add a GType definition and make it have the same size/signature as the C GType.
Previously we were using unsigned long to represent GType, but it
seems this is not entirely correct, since GType is typedefed to
gsize if sizeof(gsize) != sizeof(gulong). Here we define gzise as
QIntegerForSizeof<size_t>::Unsigned, as gzise is also typedefed to
an unsigned type of the same size as size_t (from glib's configure.ac).