VideoWidget: add support for watching a pipeline and drop the broken support for autoplug elements.
The previous support for autoplug elements was broken, as for example
an autovideosink on a playbin2 does not create the child element on the
main thread, which causes trouble for us.
The new method of watching the pipeline will work nicely on most pipelines,
like playbin2, and follows gstreamer's documentation to the letter, which
guarantees it will work correctly, as long as there is only one sink on the pipeline.