Fix Structure::fromString(const QString &).
Being in a hurry when I wrote this, I forgot to add the class qualifier
in front of the function name. As a result, if this function was used
as QGst::Structure::fromString() it would generate a compiler error due to
being undefined, and if it was used as QGst::fromString() it would generate
code that would hang calling itself recursively. Thanks to this, although
this commit is strictly braking API (but not ABI), it is not going to
affect any application, since nobody can be using this function.