qtvideosink: Make the "glcontext" property write-only and available even when compiled without OpenGL support
Client code should never read and use this property. This is just used to query
OpenGL for supported features during initialization. And when the element is compiled
without OpenGL support, is is safe to let the client code set this property; we
just need to warn the user about it.