qtvideosink: assert if paint is called with a different active QGLContext than the one on the glcontext property
QPainter internally will always make the context active when it is initialized
to paint on GL surfaces (this is done in QGLPaintDevice::beginPaint which is
called from QOpenGLPaintEngine::begin). As a consequence, we can always
assert on the fact that the current context is the one of the surface that
we were initialized to draw on when the glcontext property was set.
If it is not, then this means that the given QPainter is not initialized
to paint on the same surface - bug in the client code.