qtvideosink: Make a separate qtglvideosink element that has GL capabilities.
This is to make it more clear to the user when GL is used and when it is not,
as well as prevent situations where the user would like to use GL but
initializes the glcontext property at a wrong time and gets software
rendering instead.

So, this new qtglvideosink element supports the GLSL/ArbFp backends and fails
to become READY if none of them can be initialized, while the qtvideosink
element now supports only the raster backend, since there is no way to set
a glcontext to it.

This also has the advantage that the supported caps no longer depend on when
you will set the glcontext property and the to-be-added ColorBalance interface
will only exist in qtglvideosink, since it is not supported on the raster
backend.