qtvideosink: let BufferEvent keep its own reference on the GstBuffer
Although the previous approach also looks correct, in fact there is one
case where handing over the reference can cause a memory leak. This is
the case where a buffer event is posted on the event loop, but the
sink is destroyed before this event is delivered. In this case, the
event will be discarded by the event dispatcher, but the GstBuffer
will never be unreferenced. This patch ensures that this will never
happen.