src/QGst/buffer.* add Buffer::map and ::unmap
This is needed to fill in the MapInfo structure that provides
direct access to the chunks of memory held by the buffer that
contains the data of interest.

Unfortunately I mirrored the C api of needing to map and unmap
the structure, and I'm pretty sure C++ programmers would prefer
some mechanism to unmap the MapInfo structure when it leaves scope.

Question: how to provide a better API to the mapped memory?