qtquick2videosink: fix wrong uniform value set for GST_VIDEO_FORMAT_I420 in VideoMaterialShader::updateState()
Inside the VideoMaterialShader::updateState() function the m_id_rgbTexture var
is used to set the uniform value for RGB or YUV textures. If the video format
is GST_VIDEO_FORMAT_I420 the m_id_rgbTexture var is -1. Because of this the
uniform value for RGB textures is set. This causes the incorrect colors of the
video. This patch fixes this bug.

https://bugzilla.gnome.org/show_bug.cgi?id=719439