qtvideosink: rework the way the buffer format is propagated from set_caps down to the painter
In 0.10 the format was part of the buffer, so every buffer was associated by
its format and we could pass it down to the painter with 1-1 match, meaning
that there was no possibility of interpreting an old buffer according to
the new buffer format.

In 1.0 the format is not part of the buffer, because it's not needed actually
- it rarely changes - but there is still chance of it changing while we are
playing, so we need to make sure that we don't accidentally interpret an
old buffer with a new format. So, instead of setting it in a member of
BaseDelegate from the streaming thread using a mutex, we send it to BaseDelegate
using an event, which will be enqueued in the event loop and therefore it
will be delivered in the same order as the streaming thread calls
show_frame() with a buffer that has the old format, set_caps() and
show_frame() later with a buffer that has the new format. In the previous
implementation, the mutex could not guarantee the order of the calls.