QGst/enums: Do not register StreamVolumeFormat with the type system
For some reason the 1.0 version of this enum doesn't have a GType, so
it cannot be used with the type system. Registering it with G_TYPE_ENUM
has no point because you cannot initialize a GValue to hold a G_TYPE_ENUM
for example. It's better to fail at compilation time if something
tries to use its non-existent GType.