QGst::Buffer: replace peekMemory() with getMemory() and also add memoryCount()
The peek prefix is a bad choice because in C it means getting an object
without increasing its reference count. However in C++ we cannot do
that due to smart pointers, so it's a bit confusing to use it.

Obviously for getMemory() to be useful, we also need a memoryCount() method
that returns how many memory objects there are in the buffer...